pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-eval-harness"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "duckdb>=1.0.0",
    "pyarrow>=16.0",
    "polars>=0.20",
    "plotly>=5.0",
    "streamlit>=1.35",
    "datasets>=2.0",
    "scipy>=1.12",
    "tenacity>=8.0",
    "rich>=13.0",
    "typer>=0.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-mock>=3.12", "ruff>=0.4"]

[project.scripts]
llm-eval = "llm_eval_harness.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
